Automatic logout, userview-reset & neural-networkout panel
Upon reaching the end of a user's batch the user is now automatically logged out and userview is updated back to 1.

With new script double_users we create 2N users. For the first N users the probability output of the neural network is not shown, for the remaining N users it is shown. Users X and X+N have the exact same batches as vieworders.